`GithubConfig` is moved one level up to the `github4s` package
